Working on DataLogger, KnowledgeManager, TitleScene
Working on the data logger and the knowledge manager. The data logger will be used to track the game time. The knowledge stage still needs the rest of its buttons and the generation of statements.

The TitleManager and TitleUI code has been copied over and modified from another project. It hasn't been set up yet.