Starting and Finish Knowledge Stages from World Scene Works
The knowledge stage can now be started and finished when going from the world scene. Returning to the world scene now works, as does saving data from the completed knowledge stage.

The knowledge scene doesn't work properly if you have multiple resource buttons of the same type, but that situation won't be encountered in the final game, so it should be fine.