Rough Knowledge Points System + Created Game Events
Added a basic points system for the knowledge stages, but it still needs to b expanded.

Created event scripts that are tied into certain points of progress. The game now goes to the results scene if all stages are complete, but no data is being properly sent to and interpreted in the results scene.

Probably going to start work on the action scene soon, since the major components of the world scene and knowledge scene are done. Still need to get the menus up for the gameplay scenes though.