Action Stage End Trigger and Enemy Energy Bar
The enemies can now trigger the action stage to end by reaching the end of the stage. This checks for every enemy unit for every frame, so you may change the way you do this.

The enemy's energy bar is now on-screen, but nothing has been done with it aside from writing the code to make it track the enemy's energy.

Various prefabs have received small adjustments.