Added Metal Tile + Adjusted Action Stage End Trigger and Enemy Movements
A metal tile has been added, which is where lane blasters will be placed.

Adjusted how the end of the stage is triggered by enemies. Now, enemies only check for the map boundary once they collide with a metal tile.

The enemy now uses a rigidbody for movement instead of collider.