Adjusted Day-Night Cycle Display and Timer
The game now has a simple indicator to show the transition from day to night. It's not permanent. It's just there to create a framework for day-night transitions.

The game now supports the ability for the day to roll over and transition from night to day. However, the game will likely not use this feature.