Some Optimization to Wind Indicator
The wind indicator has been optimized a bit so that it doesn't have its visuals changed every frame. It could be optimized more, but it's fine for now.

Currently there's no way for a stage to save its wind speeds. A stage struct should be made, and said struct should contain the map and wind information.