Stage Wind Setup + Partial Set Up Wind and Wave Generators
The wind and wave generators have been set up. They now generate energy based on the wind speed, and only go on their appropriate tiles. However, the wind turbines need to be able to go in the water on certain tiles close to land. That part still needs to be done.